moved methods out of model should not be part of model
