+++ empress/tree_with_webgl.html
@@ -169,9 +169,7 @@
 
 
     let numUserSelects = 0;
-    // const CLEAR_COLOR_HEX = "C4C4C4";
     const CLEAR_COLOR_HEX = "FFFFFF";
-    // const CLEAR_COLOR = 0.7686274509803922;
     const CLEAR_COLOR = 1;
     const DEFUALT_BRANCH_COLOR = "000000";
     let camera = {};
